Cleaned up code base...
