+++ old/src/cp.c
